requirements.txt
django-filter==2.3.0
django-cors-headers==3.5.0
python_http_client==3.3.1
Django==3.1.1
djangorestframework_simplejwt==4.4.0
djangorestframework==3.11.1
python-dotenv==0.14.0
psycopg2-binary==2.8.4
uwsgi>=2.0.18,<2.1
pandas==1.1.5
